오후 05:54 매장 등록 완료, 매장 조회중

+++ yedam_finalProject/src/main/java/com/yedam/finalPrj/store/service/Store.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class Store {
-	private int store_no;
+	private long store_no;
 	private String category;
 	private String name;
 	private String store_cat;
